requirements.txt
# Core FastAPI for services
fastapi
uvicorn[standard]

# LLM SDK
google-generativeai
python-dotenv

# RAG / Vector Store
chromadb
langchain
sentence-transformers

# For making requests between services
requests
langchain-community
<<<<<<< HEAD
pypdf
langchain-huggingface
langchain-chroma
=======
pypdf
>>>>>>> e0c4129 (Save my local changes)
